feat(calidad): enchufar el circuito de preguntas mal planteadas
Estaba entero en el servidor y sin una sola pantalla: el alumno podía avisar de
una pregunta ambigua, el profesorado ver su dificultad real junto a esos avisos
y anularla, y al anularla se recalificaban los exámenes ya entregados. Código
terminado, probado y sin ningún botón desde el que llegar a él.

 · El alumno avisa desde la revisión de su examen, eligiendo motivo —ambigua,
   mal redactada, respuesta incorrecta— y contando qué vio. Quien acaba de
   responder es quien mejor detecta que una pregunta está mal.
 · El profesorado lo ve en la pantalla del examen, ordenado por avisos y luego
   por dificultad, que es como se busca una pregunta mala. Al anular, se dice
   cuántos exámenes se han recalificado; es reversible y nadie baja de nota.

Sin el extremo del alumno, el panel solo podía guiarse por el porcentaje de
acierto, que no distingue una pregunta mala de un tema mal explicado. La propia
pantalla lo advierte antes de invitar a anular nada.

De paso, los recuentos se devuelven como enteros: Postgres los da como texto y
la comparación «"0" < 30» solo funcionaba por coerción de JavaScript.

Verificado el circuito entero contra la base de datos: el alumno saca un 75 %,
avisa, el profesor ve el aviso con su motivo y la dificultad, anula, la nota
sube a 100 %, reincorpora la pregunta y vuelve a 75 %.

Además, índices que faltaban en rutas transitadas (migración 066) y uno que
sobraba: idx_answerlog_user era prefijo exacto de otro, y answer_log es la tabla
que más se escribe de la plataforma, así que era una escritura de más por cada
respuesta sin ganar ningún plan de consulta.

+++ backend/src/controllers/questionQuality.controller.ts
@@ -68,12 +68,16 @@ export async function examQuality(req: Request, res: Response): Promise<void> {
          FROM respuestas r JOIN exam_questions eq ON eq.id = r.question_id
         GROUP BY r.question_id
      )
+     -- Los recuentos se convierten a entero aquí: Postgres devuelve COUNT y
+     -- ROUND como texto, y comparar "0" < 30 en la pantalla solo funciona por
+     -- coerción de JavaScript. Un contrato que se apoya en eso se rompe el día
+     -- que alguien escribe una comparación estricta.
      SELECT eq.id, eq.text, eq.format, eq.excluded_from_grading,
-            COALESCE(s.respondida, 0) AS respondida,
-            COALESCE(s.aciertos, 0) AS aciertos,
+            COALESCE(s.respondida, 0)::int AS respondida,
+            COALESCE(s.aciertos, 0)::int AS aciertos,
             CASE WHEN COALESCE(s.respondida,0) > 0
-                 THEN ROUND(100.0 * COALESCE(s.aciertos,0) / s.respondida) END AS acierto_pct,
-            (SELECT COUNT(*) FROM question_reports qr WHERE qr.exam_question_id = eq.id AND qr.kind <> 'ok') AS avisos,
+                 THEN ROUND(100.0 * COALESCE(s.aciertos,0) / s.respondida)::int END AS acierto_pct,
+            (SELECT COUNT(*) FROM question_reports qr WHERE qr.exam_question_id = eq.id AND qr.kind <> 'ok')::int AS avisos,
             (SELECT COALESCE(json_agg(json_build_object('kind', qr.kind, 'comment', qr.comment)), '[]')
                FROM question_reports qr WHERE qr.exam_question_id = eq.id AND qr.kind <> 'ok') AS detalle_avisos
        FROM exam_questions eq
